Move imported timeseries function to Stats.Imported
